пока временно поменяла все на int, потом верну обратно; добавила empty

+++ homework_01/task_02/src/utils.hpp
@@ -4,4 +4,4 @@
 #include <string_view>
 #include <vector>
 
-double Calculate(const std::string& data);
+int Calculate(const std::string& data);
